requirements.txt
openai==1.13.3
pyyaml==6.0.1
